finalized buffer data structure and read operation such that it can read files longer that 32-bit values can hold

+++ src/Core/Filesystem/Filesystem.h
@@ -21,7 +21,7 @@ namespace Kraid
         bool Write(const uint8* data, uint64 size);
         bool Append(const uint8* data, uint64 size);
         Buffer Read();
-        uint32 GetSize() const;
+        uint64 GetSize() const;
 
         ~File();
